Renombra los ejemplos de sapi_examples para mantener el mismo estilo que los originales.
