+++ sapi_examples/edu-ciaa-nxp/clock_01_cycles_counter_demo/src/cycles_counter_demo.c
