+++ sapi_examples/edu-ciaa-nxp/pwm_02_rgb_controller_uart/Makefile
