+++ sapi_examples/edu-ciaa-nxp/stdio_01_printf_sprintf/Makefile
